requirements.txt
torch==1.5.0
torchvision==0.6.0
tensorboard==2.0.0
prefetch-generator==1.0.1
apex==0.1
Glymur==0.9.1
imgaug==0.4.0
matplotlib==3.2.1
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.1
